Added new functions and updated transparent
Added functions 'store' and 'create_tuple'. 'store' allows the transparent func. to continue running whilst keeping the variables that it came up with. 'create_tuple' forms the necessary tuples from those values that were stored and returns an array of them to be used. The main body of transparent has been changed so when it reaches the endpoint of an image, it will start to crop the images.